blogBackend/requirements.txt
asgiref==3.9.1
bleach==5.0.1
dj-database-url==3.0.1
Django==5.2.4
django-bleach==3.1.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
tinycss2==1.1.1
typing_extensions==4.14.1
webencodings==0.5.1
whitenoise==6.9.0
